:new: Standardize context API for components
